+++ Misc/NEWS.d/next/Library/2017-09-28-12-01-56.bpo-11063.neO7aq.rst
@@ -1,0 +1,3 @@
+uuid module: Move getnode() and uuid1() implementations to a new _uuid1
+submodule which is now loaded on demand to reduce side effects on "import
+uuid".
